+++ RenderSystems/Metal/src/OgreMetalStagingTexture.mm
@@ -113,7 +113,7 @@ namespace Ogre
         MetalTextureGpu *dstTextureMetal = static_cast<MetalTextureGpu *>( dstTexture );
 
         const size_t srcOffset =
-            reinterpret_cast<uint8 *>( srcBox.data ) - reinterpret_cast<uint8 *>( mMappedPtr );
+            size_t( reinterpret_cast<uint8 *>( srcBox.data ) - reinterpret_cast<uint8 *>( mMappedPtr ) );
         const NSUInteger destinationSlice =
             ( dstBox ? dstBox->sliceStart : 0 ) + dstTexture->getInternalSliceStart();
         NSUInteger xPos = static_cast<NSUInteger>( dstBox ? dstBox->x : 0 );
